+++ packages/langchain/lib/src/chains/base.dart
@@ -24,17 +24,17 @@ import 'models/models.dart';
 ///   execute a Chain. This takes inputs as a dictionary and returns a
 ///   dictionary output.
 /// - [run] A convenience method that takes inputs and returns the output as a
-///   string. This method can only be used for a subset of chains and cannot
-///   return as rich of an output as [call].
+///   string. This method can only be used if the Chain has a single string
+///   output.
 /// {@endtemplate}
-abstract class BaseChain {
+abstract class BaseChain<MemoryType extends BaseMemory> {
   /// {@macro base_chain}
   const BaseChain({
     this.memory,
   });
 
   /// Memory to use for this chain.
-  final BaseMemory? memory;
+  final MemoryType? memory;
 
   /// Return the string type key uniquely identifying this class of chain.
   String get chainType;
@@ -49,7 +49,10 @@ abstract class BaseChain {
   String get runOutputKey {
     if (outputKeys.length != 1) {
       throw LangChainException(
-        message: '`run` only supports one key output key. Got $outputKeys',
+        message: '`chain.run` method can only be used with chains that return '
+            'a single string output. $chainType chain returns '
+            '${outputKeys.length} output: $outputKeys. '
+            'Use `chain.call` method instead.',
       );
     }
     return outputKeys.first;